move prices to somewhre else (conf file or something)

+++ app/controllers/subscriptions_controller.rb
@@ -20,7 +20,7 @@ class SubscriptionsController < ApplicationController
   def payment_params
     {
       name: params['name'],
-      price: params['price'].to_i,
+      price: params['price'].to_f,
       quantity: params[:quantity].to_i,
       token: params['paymill_card_token']
     }
